requirements.txt
langchain
pandas
openai
python-dotenv
langchain-openai
langchain-experimental
